overworld created
So I made an overworld level with really bad placeholder art. The
interesting thing to note is that the controller of the world
automatically draws the lines connecting each world's indicator dot
together. This will automatically work with as many levels as you feel
like adding.

I haven't thought of a good way to choose an order, at the moment, it
just draws the lines in the order of nearest to farthest from the
overworld controller object.